Add files via upload
There are three main images to note here one of them being the zenmap screenshot with the open ports and list of active devices. The ip address and mac addresses have been censored to protect privacy. The other two screenshots are of wireshark list of captured packets as well as a screenshot of a singular captured packet on a virtual machine while nmap scan was running.

From the zenmap scan it can be seen that the most common exploitable open ports are port numbers 135, 139 ,445 and 1900 which are all TCP which was obtained from the tcp scan.

port 135- Runs Microsoft remote procedure call. This can be used to execute remote code executions allowing for lateral movement
port 139 - Runs NETBIOS-SSN and can be used for SMB relay attacks.
port 445 - Microsoft-DS (SMB) 
port 1900- Universal Plug and Play (UPnP) it is primarily used in DDoS amplification due to misconfigurations on IoT devices.